requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
streamlit>=1.0.0
numpy>=1.19.5
mediapipe>=0.8.9
scikit-learn>=0.24.2
pillow>=8.3.1
tqdm>=4.62.2
matplotlib>=3.4.3
pandas>=1.3.2 